Secrets Service Keyring Tool
This tool provides a way to securely store and retrieve passwords for specific services and users. To use the tool,
run the script with the appropriate SECURITY_MODE value. The 'get' mode will retrieve the password for the given
service and user from the Keyring, while the 'set' mode will securely store a password for the given service and
user in the Keyring.